+++ openproblems/tasks/dimensionality_reduction/api.py
@@ -1,7 +1,6 @@
 from ...data.sample import load_sample_data
 from ...tools.decorators import dataset
 from ...tools.normalize import log_cp10k
-from . import _utils
 
 import numpy as np
 
@@ -32,7 +31,6 @@ def sample_dataset():
     adata = load_sample_data()
     adata = log_cp10k(adata)
     adata.uns["n_genes"] = adata.shape[1]
-    adata.obsm["X_ranking"] = _utils.ranking_matrix(adata.X)
     return adata
