+++ openproblems/tasks/dimensionality_reduction/datasets/mouse_hspc_nestorowa2016.py
@@ -1,7 +1,6 @@
 from ....data.mouse_hspc_nestorowa2016 import load_mouse_hspc_nestorowa2016
 from ....tools.decorators import dataset
 from ....tools.normalize import log_cp10k
-from .._utils import ranking_matrix
 
 
 @dataset(
@@ -17,5 +16,4 @@ def mouse_hspc_nestorowa2016(test=False):
     adata = load_mouse_hspc_nestorowa2016(test=test)
     adata.uns["n_genes"] = adata.shape[1]
     adata = log_cp10k(adata)
-    adata.obsm["X_ranking"] = ranking_matrix(adata.X)
     return adata
