+++ openproblems/tasks/dimensionality_reduction/datasets/zebrafish.py
@@ -1,7 +1,6 @@
 from ....data.zebrafish import load_zebrafish
 from ....tools.decorators import dataset
 from ....tools.normalize import log_cp10k
-from .._utils import ranking_matrix
 
 
 @dataset(
@@ -16,13 +15,7 @@ from .._utils import ranking_matrix
     ),
 )
 def zebrafish_labs(test=False):
-    import scanpy as sc
-
     adata = load_zebrafish(test=test)
-    if not test:
-        # this dataset is too big
-        sc.pp.subsample(adata, n_obs=25000)
     adata.uns["n_genes"] = adata.shape[1]
     adata = log_cp10k(adata)
-    adata.obsm["X_ranking"] = ranking_matrix(adata.X)
     return adata
